HLE binding improvements

+++ rpcs3/Emu/SysCalls/SysCalls.h
@@ -66,6 +66,6 @@ class PPUThread;
 class SysCalls
 {
 public:
-	static void DoSyscall(PPUThread& CPU, u32 code);
+	static void DoSyscall(PPUThread& CPU, u64 code);
 	static std::string GetHLEFuncName(const u32 fid);
 };
